requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
faiss-cpu
pydantic
python-dotenv
requests
